% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getcbs.R
\name{getcbs}
\alias{getcbs}
\title{download data from CBS}
\usage{
getcbs(id, fname, toc = cbstoc, base_url = NA)
}
\arguments{
\item{id}{CBS identifier for the data, see Identifier field in TOC}

\item{toc}{name of cbstoc object}

\item{name}{name of file to store CBS data}

\item{datadir}{directory to store data}
}
\value{
A list with several objects: data contains the data.frame
with data, meta contains the meta data, tocentry contains the meta
data from the TOC, keyName is the translation table, and fieldnames
is a table with filednames and grouping id's.
}
\description{
Download data and meta data from CBS
}
\details{
This function downloads the data and the meta data from a CBS
table and stores it to disk. It checks if the table exists using the TOC and then checks
if the file already exists. If not, the data is downloaded. If the
file is allready present, then the data is read from the file.

After downloading the fieldnames of the data are renamed and the
unit of the field is added to the fieldname. A translation table is
created with the original fieldname and the fieldname created by
this function.

CBS fieldnames folow a certain hierarchilcal grouping order. For
example in the data about agriculture you find the groups farms and
animals. In each group there is the variable pigs_amount
(varkens_aantallen). In the farm group this is refering to the
amount of farms with pigs, in the animal group this is refering to
the amount of pigs. From the fieldname it's not clear to which
group the field belongs, use the function \code{getFieldParents} to
get this information.

downloaded files are stored in a directory called ./data
}
\examples{


}
