% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getcbstoc.R
\name{getcbstoc}
\alias{getcbstoc}
\title{Get CBS table of contents}
\usage{
getcbstoc(fname, base_url = NA)
}
\arguments{
\item{fname}{file name of the toc file}
}
\value{
A list with the CBS toc
}
\description{
Download the table of contents from the CBS API
}
\details{
This function downloads the table of contents from the CBS website
and stores is into a file. If the file exists, the data is not
reloaded from the CBS, reducing network overhead.

downloaded files are stored in a directory called ./data
}
\examples{
cbstoc <- getcbstoc()

}
