% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFieldParents.R
\name{getFieldParents}
\alias{getFieldParents}
\title{Get grouping info}
\usage{
getFieldParents(fieldn, cbsobj)
}
\arguments{
\item{fieldn}{field name}

\item{cbsobj}{object which was returned by getcbs}
}
\value{
a vector with the hierarchical groups, starting with the
top group
}
\description{
retrieve the hierarchical groupings for a getcbs object
}
\details{
CBS fieldnames folow a certain hierarchical grouping order. For
example in the data about agriculture you find the groups farms and
animals. In each group there is the variable pigs_amount
(varkens_aantallen). In the farm group this is refering to the
amount of farms with pigs, in the animal group this is refering to
the amount of pigs. From the fieldname it's not clear to which
group the field belongs, This function get's this information.
}
\examples{
cbstoc <- getcbstoc()

}
